created CuisinePages
this page will be dynamically used to display the multiple cuisine pages

and also linked it with the home page
