+++ src/pages/Home.jsx
@@ -6,11 +6,13 @@ import "../styles/Cards.css";
 import CuisineCards from '../components/CuisineCards'
 import Footer from "../components/Footer";
 import Header from "../components/Header";
+import Search from "../components/Search";
 
 const Home = () => {
   return (
     <div>
       <Header/>
+      <Search/>
       <CuisineCards/>
       <Footer/>
     </div>
