+++ src/styles/nav.css
@@ -88,21 +88,6 @@ nav a {
 }
 
 
-/* .user-icon {
-    width: 40px;
-    height: 40px;
-    font-weight: bold;
-    border-radius: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-left: 10px;
-    font-size: 18px;
-    background: black;
-    color: white; 
-    font-family: 'Arvo', serif;
-} */
-
 .logout-btn {
     position: relative; /* Ensure underline works */
 }
